Externally controlled (env var) skip of temp file deletion to assist debugging
bzr-revno: 283.2.1

+++ tests/vmtests/__init__.py
@@ -134,8 +134,13 @@ class TempDir:
                               stdout=DEVNULL, stderr=subprocess.STDOUT)
 
     def __del__(self):
-        # remove tempdir
-        shutil.rmtree(self.tmpdir)
+        if (os.environ.get('KEEP_VMTEST_DATA') is not None):
+            # remove tempdir
+            logger.debug('deleting tmp files')
+            shutil.rmtree(self.tmpdir)
+        else:
+            logger.debug('Found Environment Variable KEEP_VMTEST_DATA')
+            logger.debug('not deleting tmp files')
 
 
 class VMBaseClass:
